requirements.txt
Pillow
yacs
tqdm
requests